refactor: extract helpers/streams.py (Phase 6.3)
Move the bounded SSE slot state machine out of dashboard.py's module-level
bag into helpers/streams.py. Self-contained module — no external reader
of the counters, so all state lives here cleanly:

  SSE_MAX_SECONDS                  — wall-clock cap for any SSE stream
  MAX_LOG_STREAM_CLIENTS           — 10
  MAX_HEALTH_STREAM_CLIENTS        — 10
  MAX_BRAIN_STREAM_CLIENTS         — 5
  _stream_clients_lock             — threading.Lock
  _active_*_stream_clients         — per-kind counters (3)
  _acquire_stream_slot(kind)       — reserve a slot; False if cap reached
  _release_stream_slot(kind)       — return a slot; idempotent floor at 0

Call sites (all via `_d.<name>`): routes/brain.py (brain-stream),
routes/health.py (health-stream), routes/infra.py (logs-stream).

## Notes

dashboard.py had TWO identical copies of the constants + state (lines
243-250 and 6115-6122) from the embedded-HTML-string quirk. Both removed.

## E2E

- Unit tests: acquire cap=5 for brain, 6th returns False, release all,
  re-acquire works, unknown kind returns False
- 39/39 API endpoints return 200
- 3/3 SSE streams open cleanly: /api/brain-stream, /api/health-stream,
  /api/logs-stream all return HTTP/1.1 200 OK on HEAD
- Zero tracebacks

## Sizes

  dashboard.py       : -58 lines (2 copies of 7-line state block +
                                  2 copies of SSE_MAX_SECONDS + two
                                  ~18-line function bodies removed;
                                  5-line re-export import added)
  helpers/streams.py : new (72 lines with docstrings)
